add move participant in analogy to delete
