+++ org.eclipse.ui.views.pdf/src/org/eclipse/ui/views/pdf/refactoring/DeletePdfParticipant.java
@@ -1,0 +1,33 @@
+package org.eclipse.ui.views.pdf.refactoring;
+
+import org.eclipse.core.runtime.CoreException;
+import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.core.runtime.OperationCanceledException;
+import org.eclipse.ltk.core.refactoring.Change;
+import org.eclipse.ltk.core.refactoring.RefactoringStatus;
+import org.eclipse.ltk.core.refactoring.participants.CheckConditionsContext;
+import org.eclipse.ltk.core.refactoring.participants.DeleteParticipant;
+
+public class DeletePdfParticipant extends DeleteParticipant {
+
+	@Override
+	protected boolean initialize(Object element) {
+		new PdfInScoreViewCloser(element);
+		return false;
+	}
+
+	@Override
+	public String getName() {
+		return "Close Score View";
+	}
+
+	@Override
+	public RefactoringStatus checkConditions(IProgressMonitor pm, CheckConditionsContext context) throws OperationCanceledException {
+		return new RefactoringStatus();
+	}
+
+	@Override
+	public Change createChange(IProgressMonitor pm) throws CoreException, OperationCanceledException {
+		return null;
+	}
+}
